исправил замечания по прошлому заданию
